+++ docs/en/changes/changes.md
@@ -27,6 +27,7 @@
 * Support collecting memory buff/cache metrics in VM monitoring.
 * PromQL: Remove empty values from the query result, fix `/api/v1/metadata` param `limit` could cause out of bound.
 * Support monitoring the total number metrics of k8s StatefulSet and DaemonSet.
+* Support Amazon API Gateway monitoring.
 
 #### UI
 * Revert: cpm5d function. This feature is cancelled from backend.